### 2019-07-25 #330 @mpadge
mapscanner
Submitting Author: mark padgham (@mpadge)
Repository:  [mpadge/mapscanner](https://github.com/mpadge/mapscanner)
Version submitted: 0.0.1
Version accepted: TBD

---

```
Package: mapscanner
Title: Print Maps, Draw on Them, Scan Them Back In
Version: 0.0.1
    person("Mark", "Padgham", email="mark.padgham@email.com", role=c("aut", "cre")),
    person("Michael D.", "Sumner", role = "aut", email = "mdsumner@gmail.com"),
    person("Stanislaw", "Adaszewski", role = "cph",
        comment = "author of include concaveman-cpp code"))
Description: Print maps, draw on them, scan them back in, and convert to spatial
    objects.
License: GPL-3 | BSD_2_clause + file LICENSE
NeedsCompilation: yes
    cli,
    curl,
    fs,
    glue,
    jpeg,
    magick,
    memoise,
    pdftools,
    png,
    raster,
    Rcpp,
    reproj,
    RNiftyReg,
    sf,
    slippymath,
    RTriangle,
    gibble,
    polyclip,
    unjoin,
    rlang
    lwgeom,
    mapview,
    osmdata,
    spelling
LinkingTo: Rcpp
SystemRequirements:
    C++11
Language: en-GB
```

[`mapscanner`](https://github.com/mpadge/mapscanner) fits nicely into both data extraction and geospatial data. It is definitely geospatial because it concerns maps, and ultimately generates spatial output in standard (sf) format. It is nevertheless primarily a "data extraction" tool that extracts useful (spatially) structured data from unstructured input in either pdf or image formats.

The target audience comprises social researchers interested in spatial phenomena such as subjective perceptions of regions or areas, or perceptions of location. More specifically, these kinds of analyses are currently restricted to (highly) commercial service provision, and the package aims to aid and empower social researchers in financially disadvantageous circumstances by providing analytic abilities that would otherwise be prohibitively expensive.

No R packages whatsoever that we are aware of. Vaguely similar software exists in other languages, such as [`mapmatcher`](https://github.com/julienperret/mapmatcher) (in `java`; long abandoned) and [`Maplat`](https://github.com/code4nara/Maplat) (`.js`; mostly in Japanese, and comparable only in algorithmic terms; purpose is entirely different).

Pre-submission issue #322, approved for submission by @noamross :fireworks:

  - [x] The package is deposited in a long-term repository with the DOI: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3647922.svg)](https://doi.org/10.5281/zenodo.3647922)
