+++ inst/paper.md
@@ -1,0 +1,35 @@
+---
+title: mapscanner
+tags:
+    - maps
+    - spatial
+    - R
+authors:
+    - name: Mark Padgham
+      affiliation: 1
+    - name: Michael D. Sumner
+      affiliation: 2
+affiliations:
+    - name: Active Transport Futures, Muenster, Germany
+      index: 1
+    - name: Australian Antarctic Division, Hobart, Australia.
+      index: 2
+date: 25 July 2019
+bibliography: vignettes/mapscanner.bib
+nocite: |
+  @*
+---
+
+# Summary
+
+`mapscanner` is an **R** package that enables maps to be printed out, drawn on,
+scanned back in, and have the drawn components converted to spatial objects. It
+is intended for use in social surveys and similar endeavours in which printed
+paper maps are preferred over digital screens, and enables marks drawn by hand
+on paper maps to be converted to spatial objects. The use of paper maps is
+often advantageous because (i) they cost much less than computer devices
+otherwise needed for digital equivalents; and (ii) they may in many situations
+be deemed to yield more reliable or accurate results through removing the
+psychological barriers often associated with screen-based surveys [@Toninelli2016].
+
+# References
