+++ vignettes/mapscanner.bib
@@ -1,0 +1,82 @@
+@article{Antoun2017,
+    author = {Antoun, Christopher and Couper, Mick P. and Conrad, Frederick G.},
+    title = "{Effects of Mobile versus PC Web on Survey Response Quality: A Crossover Experiment in a Probability Web Panel}",
+    journal = {Public Opinion Quarterly},
+    volume = {81},
+    number = {S1},
+    pages = {280-306},
+    year = {2017},
+    month = {03},
+    abstract = "{Survey participants are increasingly responding to Web surveys on their smartphones as opposed to their personal computers (PCs), and this change brings with it some potential data-quality issues. This study reports on a randomized crossover experiment to compare the effect of two different devices, smartphones and PCs, on response quality in a Web survey conducted in a probability-based panel. Participants (n = 1,390) were invited to complete an online questionnaire on both a smartphone (mobile Web) and PC (PC Web) in sequence. We hypothesized that smartphone use would result in lower-quality responses because people are more likely to use smartphones while multitasking or while around other people and because they could have difficulty recording their answers using a small touchscreen. While we found that respondents who participated in this study were more likely to multitask and more likely to be around other people when using smartphones, these factors had little impact on data quality. Respondents were at least as likely to provide conscientious and thoughtful answers and to disclose sensitive information on smartphones as on PCs. When using smartphones, however, respondents seemed to have trouble accurately moving a small-sized slider handle and a date-picker wheel to the intended values. Overall, we find that people using smartphones can provide high-quality responses, even when their context is more distracting, as long as they are presented with question formats that are easy to use on small touchscreens.}",
+    issn = {0033-362X},
+    doi = {10.1093/poq/nfw088},
+    url = {https://doi.org/10.1093/poq/nfw088},
+    eprint = {http://oup.prod.sis.lan/poq/article-pdf/81/S1/280/13942296/nfw088.pdf},
+}
+
+@article{Toninelli2016,
+    author = {Toninelli, Daniele and Revilla, Melanie},
+    title = "{Smartphones vs PCs: Does the Device Affect the Web Survey
+        Experience and the Measurement Error for Sensitive Topics? - A
+        Replication of the Mavletova & Couper’s 2013 Experiment}",
+    journal = {Survey Research Methods},
+    volume = {10},
+    number = {2},
+    pages = {153--169},
+    year = {2016},
+    abstract = "{More and more respondents use mobile devices to complete web
+        surveys. These devices have different characteristics, if compared to
+        PCs (e.g. smaller screen sizes and higher portability). These
+        characteristics can affect the survey responses, mostly when a
+        questionnaire includes sensitive questions. This topic was already
+        studied by Mavletova and Couper (2013), through a two-wave experiment
+        comparing PCs and mobile devices results for the same respondents in a
+        Russian opt-in panel. We replicated this cross-over design, focusing on
+        an opt-in panel for Spain, involving 1,800 panellists and comparing PCs
+        and smartphones. Our results support most of Mavletova and Couper’s
+        (2013) findings (e.g. generally the used device does not significantly
+        affect the reporting of sensitive information), confirming their
+        robustness over the two studied countries. For other results (e.g. trust
+        in data confidentiality), we found differences that can be justified by
+        the diverse context/culture or by the quick changes that are still
+        characterizing the mobile web survey participation.}",
+    doi = {10.18148/srm/2016.v10i2.6274 },
+    url = {https://ojs.ub.uni-konstanz.de/srm/article/view/6274},
+}
+
+@article{Wright1998,
+    ISSN = {0033362X, 15375331},
+    URL = {http://www.jstor.org/stable/2749663},
+    abstract = {This study compares estimates of self-reported substance use
+        and psychological well-being obtained in computer-assisted and
[…]
