### 2019-07-30 4257885
add styles param to generate_map; closes #10
