### 2019-08-01 #330 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @mpadge !

Here's the output from [goodpractice](https://github.com/MangoTheCat/goodpractice/). You don't need to address these now, it's info for reviewers to use to get started.

- the long code line seems fine
- i think this may be a case of FALSE positives (get it?) for the T/F check in goodpractice, so ignore that
- I had trouble installing locally. Seems it does install form github okay, but then locally I get an openmp error, the error suggested settting an env var, so `Rscript -e 'Sys.setenv(KMP_DUPLICATE_LIB_OK=TRUE); library(devtools); document(); install()'` worked

```r
── GP mapscanner ─────
It is good practice to

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    R/RcppExports.R:19:1

  ✖ avoid 'T' and 'F', as they are just variables which are set to the logicals 'TRUE' and 'FALSE' by default, but are not reserved words and hence can
    be overwritten by the user.  Hence, one should always use 'TRUE' and 'FALSE' for the logicals.

    R/aggregate_polys.R:NA:NA
    R/aggregate_polys.R:NA:NA
 ───────────────
```

Seeking reviewers now 🕐

---

Reviewers:

- xx
- xx
