### 2019-08-02 86090e9
add rOS review link to readme
