### 2019-08-23 #330 @potterzot
In just under the deadline! @mpadge thank you for your submission, this is a super cool package. We may end up using it in some work we're doing with water rights, so it's great to see! That said, I couldn't actually get it to download maps given a `bbox`, but looking forward to trying it out once I can.

I printed a USGS topo map and draw a trail on it to check it's application for hiking and was excited with the results. While the package is explicitly aimed at researchers, another use case would be hikers and runners who want to plan a route but don't want to click on an online map. This would let you draw the route, scan it in to the map, and have it uploaded as a shape. If it could be then converted to a GPX file that would be really neat.

I've used the below checklist from the rOpensci review template. Comments at the end are my additions. Please feel free to comment with any clarifying questions.

>##### For packages co-submitting to JOSS
>
>- [X] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [X] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

The package seems great. There are some problems I had with checking and running an example, and some code style differences. I left check boxes above unchecked where there were comments I made to suggest changes. The `Statement of Need` and `Vignette` sections above are unchecked beccause the README does not have a clear statement of use, although the vignette does. I would suggest just copying that to the README, or possibly making it a bit simpler as well. The vignette is fine but was not building properly for me. The `Functionality` and `Automated Tasks` checkboxes are unchecked because I received an error when trying to run the example from the README, detailed below and I think that automated tests of the API may help with identifying possible issues when attempting to download maps. Finally, `Package Guidelines` is unchecked because of style discrepancies that are detailed below.

#### Specific Comments

##### Errors

- `devtools::check()` fails with this error when trying to build the vignette:

File mapshot-polys.png not found in resource path
```

- When I try to run the code from the README, I get the following error:

bbox <- rbind (c (-117.25000, -117.37500),
               c (45.12500, 45.16667))

ms_generate_map (bbox, max_tiles = 16L, mapname = "eagle_cap")
#> Error in curl::curl_download(url = api_query, outfile) : HTTP error 401.
```

`curl::curl_download()` works with a simple html file, so it doesn't seem to be curl that is breaking.

##### Documentation

- Since I got an error (detailed below) when attempting to generate a map, instead I printed a pdf of a map that I had, drew on it, and then scanned it in. The maps are 2.2MB and 9.9MB. The `ms_rectify_maps()` function ran for three hours and then it was getting close to the time I had to leave so I killed it. It did not exit cleanly. Providing a check interrupt as in [this stackexchange](https://stackoverflow.com/questions/40563522/r-how-to-write-interruptible-c-function-and-recover-partial-results) would make it easier to recover if it's taking too long. As it is my R session froze and I had to kill Rstudio and start over.

- This may be a GB/US difference. Throughout the package the use of the word "scan" to mean "reading an image file into R" is confusing. In the U.S., most users will take this to mean actually using a scanner. For example, the documentation for `ms_rectify_maps()` reads "Scan in two pdf or png maps, rectify them ...". But the function isn't scanning anything. This happens in other places such as the `README`. I would suggest working on package and individual function descriptions to clarify that the package deals with images that _have been scanned in_ but doesn't do any scanning. I.e. the package allows users to converted scanned images into `sf` objects, but doesn't operate the scanner. In the US "read" would probably best replace "scan".

- __README.md__: Some expansion of the text would make the package more friendly. The __usage__ section lacks an initial description of the goal of process so it is a bit confusing what it happening (The opening paragraph of the vignette does this well for eample). It's probably worth explaining what "rectification" is for the casual user. Some sentences are incomplete, e.g. "Package comes with ...". The text in the vignette and some of the function descriptions would also be nice in the README, in accordance with the [principle of multiple points of entry:

> Is the documentation (installation instructions/vignettes/examples/demos) clear and sufficient? Does it use the principle of multiple points of entry i.e. takes into account the fact that any piece of documentation may be the first encounter the user has with the package and/or the tool/data it wraps?

- Unexported functions would benefit from comments and/or documentation to explain their purpose. This helps contributors or yourself a year from now. E.g. `p_paste` and `sf_df`.

- It looks like roxygen documentation is being done in a mix of markdown and the default, e.g. [`type = "polygons"`](https://github.com/mpadge/mapscanner/blob/23ea1378ae6f168eaeccf1df5cc4e24e21ea3ae5/R/rectify-maps.R#L18-L22) doesn't render nicely in the help file. Instead of backticks, `\code{}` will render correctly. Alternatively, you can set roxygen to use markdown formatting by default. If you choose that route, then I think default roxygen styling [like  `\pkg{sf}`](https://github.com/mpadge/mapscanner/blob/23ea1378ae6f168eaeccf1df5cc4e24e21ea3ae5/R/rectify-maps.R#L30) would have to be changed.

- The note referenced in  "see Note" in the `ms_rectify_maps()` documentation should be under __Details__. Right now it's included under the __Value__ header.

- ["Automatically save"](https://github.com/mpadge/mapscanner/blob/23ea1378ae6f168eaeccf1df5cc4e24e21ea3ae5/R/generate-maps.R#L7) should be "saved".

- I was initially confused where the map was stored until I read the [second line of the `mapname` parameter documentation](https://github.com/mpadge/mapscanner/blob/23ea1378ae6f168eaeccf1df5cc4e24e21ea3ae5/R/generate-maps.R#L13-L14). It would be helpful to add to the function description that it saves the map files in the working directory or optionally in the full path name.

- The documentation example for `ms_generate_map()`](https://github.com/mpadge/mapscanner/blob/23ea1378ae6f168eaeccf1df5cc4e24e21ea3ae5/R/generate-maps.R#L26-L31) returns this error:

    Error in curl::curl_download(url = api_query, outfile) : HTTP error 401.
```

##### Code Style

- The use of a space after the function name, e.g. `library (mapscanner)` doesn't conform to the rOpensci code style, which is also Hadley's [style guide](http://r-pkgs.had.co.nz/r.html). Moreover, in some places there is a space and in others there isn't.
- Similarly, `if()` functions should include `{}` if they extend to the next line, e.g. [here](https://github.com/mpadge/mapscanner/blob/23ea1378ae6f168eaeccf1df5cc4e24e21ea3ae5/R/rectify-maps.R#L75-L76).
- This is more of a matter of preference, but for functions from packages that are used repeatedly throughout a function, I prefer `@importFrom` rather than `::`. The code is cleaner and easier to read. I would do this with `sf` and `dplyr` and possibly `purrr`.

##### Code

- [this "tell Mike"](https://github.com/mpadge/mapscanner/blob/23ea1378ae6f168eaeccf1df5cc4e24e21ea3ae5/R/aggregate_polys.R#L82) statement should be expanded into an explicit `message` or removed entirely.
