### 2019-08-24 7dd130c
Explicitly require mapbox token
