+++ NAMESPACE
@@ -4,6 +4,7 @@ export("%>%")
 export(ms_aggregate_polys)
 export(ms_generate_map)
 export(ms_rectify_maps)
+export(set_mapbox_token)
 importFrom(RNiftyReg,niftyreg)
 importFrom(Rcpp,evalCpp)
 importFrom(magrittr,"%>%")
