### 2019-08-27 13c4947
specify need for mapbox token; closes #13
