+++ vignettes/mapscanner.Rmd
@@ -101,12 +101,18 @@ to which these may be converted.
 Map generation with `mapscanner` requires a personal token or key from
 [`mapbox`](https://mapbox.com), which can be obtained by following the links
 from
-[https://docs.mapbox.com/api](https://docs.mapbox.com/api/#access-tokens-and-token-scopes). If you already have a token, the easier way to use it with `mapscanner` is to create (or edit) a file `~/.Renviron`, and insert a line,
+[https://docs.mapbox.com/api](https://docs.mapbox.com/api/#access-tokens-and-token-scopes).
+If you already have a token, the easiest way to use it with `mapscanner` is to
+create (or edit) a file `~/.Renviron`, and insert a line,
 ``` bash
 MAPBOX_TOKEN=<my_mapbox_token>
 ```
-This will then be available every time you start R. The token may be given any
-unique name that includes "mapbox" (case insensitive). Alternatively, if you wish to keep your token truly private, and only use it for your current R session, you may load `mapscanner`, and then type `set_mapbox_token(<my_mapbox_token>)`.
+This will then be available every time you start R, without any need to
+explicitly set the token each time you want to use the package. The token may
+be given any unique name that includes "mapbox" (case insensitive).
+Alternatively, if you wish to keep your token truly private, and only use it
+for your current R session, you may load `mapscanner`, and then run
+`set_mapbox_token(<my_mapbox_token>)`.
 
 
 
