+++ R/tokens.R
@@ -5,7 +5,7 @@ get_mapbox_token <- function ()
     tok <- unique (as.character (e))
     if (length (tok) != 1 | tok == "")
         stop ("Map generation requires a mapbox API key to be set with ",
-              "Sys.setenv with a name that includes either the strings "
+              "Sys.setenv with a name that includes either the strings ",
               "'mapbox' or 'mapscanner'")
     return (tok)
 }
