+++ tests/testthat/test-aggregate-polys.R
@@ -7,6 +7,8 @@ test_that ("aggregate", {
                pts <- sf::st_sf (a = 1:3, geometry = geometry)
                overlapping_polys <- sf::st_buffer (pts, 0.75)
 
+               requireNamespace ("dplyr") # produces output
+               requireNamespace ("gibble")
                expect_silent (x <- ms_aggregate_polys (overlapping_polys))
                expect_is (x, "sf")
                expect_is (x$geometry, "sfc_MULTIPOLYGON")
