+++ DESCRIPTION
@@ -16,7 +16,6 @@ Depends: R (>= 3.5.0)
 Imports:
     cli,
     curl,
-    dplyr,
     fs,
     glue,
     jpeg,
@@ -25,20 +24,21 @@ Imports:
     memoise,
     pdftools,
     png,
+    polyclip,
     purrr,
     raster,
     Rcpp,
     reproj,
+    rlang,
     RNiftyReg,
+    RTriangle,
     sf,
     slippymath,
     tibble,
-    RTriangle,
-    gibble,
-    polyclip,
-    unjoin,
-    rlang
+    unjoin
 Suggests: 
+    dplyr,
+    gibble,
     ggplot2,
     knitr,
     lwgeom,
