### 2019-09-05 3cc34a3
:wrench:aggregate_polys for consistent code style
