### 2019-09-05 c633b36
cp vignette initial statements to README
