+++ README.md
@@ -18,6 +18,19 @@ spatial objects. Currently [under
 review](https://github.com/ropensci/software-review/issues/330#event-2513283441)
 at [rOpenSci](https://ropensci.org).
 
+## What does this package do for me?
+
+`mapscanner` is an **R** package that enables maps to be printed out,
+drawn on, scanned back in, and have the drawn components converted to
+spatial objects. It is intended for use in social surveys and similar
+endeavours in which printed paper maps are preferred over digital
+screens, and enables marks drawn by hand on paper maps to be converted
+to spatial objects. The use of paper maps is often advantageous because
+(i) they cost much less than computer devices otherwise needed for
+digital equivalents; and (ii) they may in many situations be deemed to
+yield more reliable or accurate results through removing the
+psychological barriers often associated with screen-based surveys.
+
 ## installation
 
 `mapscanner` is not (yet) on CRAN. The development version can be
@@ -34,6 +47,59 @@ library (mapscanner)
 ```
 
 ## usage
+
+The package is designed to enable the following workflow:
+
+1.  Generate a map with the
+    [`ms_generate_map()`](https://mpadge.github.io/mapscanner/reference/ms_generate_map.html)
+    function, which automatically produces both `.pdf` and `.png`
+    versions;
+
+2.  Print the `.pdf` version to use as desired in any kind of survey
+    environment;
+
+3.  Draw on the map;
+
+4.  Digitise the drawn-on (from here on, “modified”) map, converting it
+    to either `.pdf` or `.png` format; and
+
+5.  Rectify the modified version against the original via the
+    [`ms_rectify_maps()`](https://mpadge.github.io/mapscanner/reference/ms_rectify_maps.html)
+    function, which distinguishes individual annotations, and converts
+    each one to a spatial object able to be analysed in any desired
+    manner.
+
+### Practical tips
+
+Unlike most software packages, `mapscanner` is intended to aid a
+*practical* workflow, and so a few practical tips may be recommended
+here to ensure best results:
+
+1.  The original digital files generated with
+    [`ms_generate_map()`](https://mpadge.github.io/mapscanner/reference/ms_generate_map.html)
+    are necessary to rectify subsequently drawn-on and scanned maps, and
+    so must be retained at all times.
+2.  Marks drawn on maps should be *coloured* – any black or grey
+    markings will be ignored. This has the advantage that individual
+    annotations *not* intended to be converted to spatial objects (such
+    as unique identification or participant codes) may be made on maps
+    in black or grey.
+3.  For drawings of areas, care should be taken to ensure all lines form
+    closed polygons; only in cases where this is not possible or
+    practicable, the `type = "hulls"` argument should be used in map
+    rectification, as described further below.
+4.  Digitised versions should contain *white* borders, so do not, for
+    example, photograph modified maps lying on dark surfaces. If maps
+    are to be photographed, then best results can be achieved by simply
+    placing them on a larger, enclosing sheet of white paper.
+
+The following two sections describe the two primary functions of the
+`mapscanner` package, corresponding to the two primary steps of
+producing maps to be used in surveys, and rectifying modified maps
+against these originals in order to extract spatial objects. The second
+of these sections also describes the kinds of markings able to be
+recognised, and the kinds of spatial objects to which these may be
+converted.
 
 ### Mapbox API tokens
 
@@ -102,7 +168,7 @@ system.time (res <- ms_rectify_maps (f_orig, f_mod, type = "polygons"))
 #> ❯ converting to spatial format 
 ✔ converting to spatial format
 #>    user  system elapsed 
-#>  50.908   0.704  13.183
+#>  39.881   0.739  11.675
 res
 #> Simple feature collection with 2 features and 0 fields
 #> geometry type:  POLYGON
