### 2019-09-05 mpadge/mapscanner#15 @mpadge
auto-reduce image files that are too large
[Review of @potterzot ](https://github.com/ropensci/software-review/issues/330#issuecomment-524492943) tried feeding in large files, which caused `RNiftyReg` to take forever and ultimately fail. That package will only accept max file dimension of 2,048 anyway, so I should auto-resize large files to smaller equivalents for rectification.
