### 2019-09-05 d4aadf0
auto-reduce size of large files; closes #15
