### 2019-09-05 5652a90
clarify use of 'scan' in docs; closes #16
