### 2019-09-05 fb7c795
explicitly specify mapname path in man entry
