+++ man/ms_generate_map.Rd
@@ -31,7 +31,9 @@ package containing all data used to generate the map.
 }
 \description{
 Generate a map image for a specified area or bounding box. Map is
-automatically saved in both \code{.pdf} and \code{.png} formats.
+automatically saved in both \code{.pdf} and \code{.png} formats, by default in current
+working directory, or alternative location when \code{mapname} includes the full
+path.
 }
 \examples{
 \dontrun{
