+++ R/generate-maps.R
@@ -3,10 +3,10 @@
 
 #' ms_generate_map
 #'
-#' Generate a map image for a specified area or bounding box. Map is
-#' automatically saved in both `.pdf` and `.png` formats, by default in current
-#' working directory, or alternative location when `mapname` includes the full
-#' path.
+#' Generate a map image for a specified area or bounding box, by downloading
+#' tiles from \url{https://mapbox.com}. Map is automatically saved in both
+#' `.pdf` and `.png` formats, by default in current working directory, or
+#' alternative location when `mapname` includes the full path.
 #'
 #' @param bbox Either a string specifying the location, or a numeric bounding
 #' box as a single vector of (xmin, ymin, xmax, ymax), or a 2-by-2 matrix with
