### 2019-09-05 73e1f08
include mapbox url in get_mapbox_token error
