### 2019-09-05 #330 @mpadge
Thanks @potterzot for the very productive and helpful review, for which I now offer the following responses, quoting your review comments throughout as appropriate. But first:

## need for mapbox API token

The previous inability to use one of the main functions was caused by my hard-coding an API token as environmental variable, as [described above](https://github.com/ropensci/software-review/issues/330#issuecomment-524509013). This has been rectified so that the package will automagically work if it detects any kind of environmental variable with "mapbox" or "mapscan" in the name, otherwise it will prompt the user with [the following](https://github.com/mpadge/mapscanner/blob/master/R/tokens.R):

```
"Map generation requires a mapbox API key to be set with Sys.setenv
with a name that includes either the strings 'mapbox' or 'mapscanner'.
Tokens can be obtained from
https://docs.mapbox.com/api/#access-tokens-and-token-scopes"
```

---

## review responses

> There are ... some code style differences.

These should have now been rectified, particularly by rendering the code style of the distinct [`aggregate_polys.R`](https://github.com/mpadge/mapscanner/blob/master/R/aggregate_polys.R) function to be consistent with the remainder of the package.

> The Statement of Need and Vignette sections above are unchecked beccause the > README does not have a clear statement of use, although the vignette does. I > would suggest just copying that to the README, or possibly making it a bit > simpler as well.

and further:

> README.md: Some expansion of the text would make the package more friendly. The usage section lacks an initial description of the goal of process so it is a bit confusing what it happening (The opening paragraph of the vignette does this well for eample). It's probably worth explaining what "rectification" is for the casual user. Some sentences are incomplete, e.g. "Package comes with ...". The text in the vignette and some of the function descriptions would also be nice in the README, in accordance with the [principle of multiple points of entry:

The initial vignette sections describing package motivation and principles of usage have been copied to the `README` as suggested.

> The vignette is fine but was not building properly for me... The > Functionality and Automated Tasks checkboxes are unchecked because I received > an error when trying to run the example from the README, detailed below

I presume these issues were because of my previous error in handling mapbox API keys, and hope that it should now build once a key has been set via appropriately named environmental variable. (`Sys.setenv ("...MAPBOX...") = <my_key>`).

> I think that automated tests of the API may help with identifying possible > issues when attempting to download maps.

While the package has 100% coverage, this is admittedly because the bit of code directly responsible for downloading the maptiles is currently [not tested](https://github.com/mpadge/mapscanner/blob/master/R/generate-maps.R#L57-L83). These lines could be mock tested, but would nevertheless require an API key that would have to be hard-coded in one particular way, and would thus still potentially not capture all possible ways by which such keys might be set and used. The new code for dealing with [tokens via environmental variables](https://github.com/mpadge/mapscanner/blob/master/R/tokens.R) *should* be failsafe for most use cases, and should issue sufficiently clear guidelines to ensure that tokens are appropriately set, as described above.

## Errors

> devtools::check() fails with this error when trying to build the vignette:
`File mapshot-polys.png not found in resource path`

> When I try to run the code from the README, I get the following error ... > curl::curl_download() works with a simple html file, so it doesn't seem to be > curl that is breaking.

The latter, and as far as I can tell, the former as well, were both caused by previous lack of appropriate, and appropriately described need for, mapbox token.

## Documentation

> I would also add more information about where the maps are being downloaded from.

Documentation for `ms_generate_map()` now states that the function:
> Generate(s) a map image for a specified area or bounding box, by downloading tiles from \url{https://mapbox.com}.

(See [this commit](https://github.com/mpadge/mapscanner/commit/001e24c465d310588941a5c4d5969a4c58a629ee)).

> Since I got an error (detailed below) when attempting to generate a map, instead I printed a pdf of a map that I had, drew on it, and then scanned it in. The maps are 2.2MB and 9.9MB. The ms_rectify_maps() function ran for three hours and then it was getting close to the time I had to leave so I killed it. It did not exit cleanly. Providing a check interrupt as in this stackexchange would make it easier to recover if it's taking too long. As it is my R session froze and I had to kill Rstudio and start over.

These issues extend from the [`RNiftyReg` package](https://github.com/jonclayden/RNiftyReg), for which I have opened [an issue](https://github.com/jonclayden/RNiftyReg/issues/8) requesting an interrupt function. There have also been previous observations that functions in this package [do not exit cleanly](https://github.com/jonclayden/RNiftyReg/issues/3), but causes and solutions unfortunately lie beyond my control in the current package.

What i have nevertheless done is to implement an automatic reduction in size of large images, as detailed in [issue #15](https://github.com/mpadge/mapscanner/issues/15). I have fixed the size to ensure files submitted to `RNiftyReg` are <1MB in size, although there could be more flexible ways to do this? Nevertheless, as noted there, `RNiftyReg` routines are generally designed to only work with images up to a maximum of 2,048 pixels in any one dimension, so this seems a safe approach. I'd welcome any further thoughts or suggestsions @potterzot ...

> This may be a GB/US difference. Throughout the package the use of the word "scan" to mean "reading an image file into R" is confusing. In the U.S., most users will take this to mean actually using a scanner. For example, the documentation for ms_rectify_maps() reads "Scan in two pdf or png maps, rectify them ...". But the function isn't scanning anything. This happens in other places such as the README. I would suggest working on package and individual function descriptions to clarify that the package deals with images that have been scanned in but doesn't do any scanning. I.e. the package allows users to converted scanned images into sf objects, but doesn't operate the scanner. In the US "read" would probably best replace "scan".

As far as I could tell, there was only that one mentioned reference to the package itself doing any "scanning". I have now modified that to say that the function can be used to, "Rectify two previously scanned-in pdf or png maps with `RNiftyReg`" (see [this issue](https://github.com/mpadge/mapscanner/issues/16)). The word "scan" is indeed used to imply actually using a scanner (or camera, or other equipment).

> Unexported functions would benefit from comments and/or documentation to explain their purpose. This helps contributors or yourself a year from now. E.g. p_paste and sf_df.

Ping co-author @mdsumner - could you please PR a fix there?

> It looks like roxygen documentation is being done in a mix of markdown and the default, e.g. type = "polygons" doesn't render nicely in the help file. Instead of backticks, \code{} will render correctly. Alternatively, you can set roxygen to use markdown formatting by default. If you choose that route, then I think default roxygen styling like \pkg{sf} would have to be changed.

Sorry, I neglected to include the essential `Roxygen: list(markdown = TRUE)` in the `DESCRIPTION`. Now fixed in [this commit](https://github.com/mpadge/mapscanner/commit/12439309659f824e01bbfe361fb7ec7767fcaa22). (And note that `\pkg` is still necessary to auto-insert the appropriate link to docs of nominated package.)

> The note referenced in "see Note" in the ms_rectify_maps() documentation should be under Details. Right now it's included under the Value header.

...

> "Automatically save" should be "saved".

Both of those fixed in [this commit](https://github.com/mpadge/mapscanner/commit/6d58ac1ea3d6a797e938f9e7ed960f903f75dea6) - thanks for noting!

> I was initially confused where the map was stored until I read the second line of the mapname parameter documentation. It would be helpful to add to the function description that it saves the map files in the working directory or optionally in the full path name.

Done in [this commit](https://github.com/mpadge/mapscanner/commit/fb7c7957394282691afbb5ef14a9b19160df1611), so that the initial description of `ms_generate_maps()` now states that a map is, "automatically saved in both `.pdf` and `.png` formats, by default in current working directory, or alternative location when `mapname` includes the full path."

## Code Style

> The use of a space after the function name, e.g. library (mapscanner) doesn't conform to the rOpensci code style, which is also Hadley's style guide.

Yeah, but it also does not *not* conform. Like i always say, "white space let's code breathe". I also explicitly mention my white space preferences in [`CONTRIBUTING.md`](https://github.com/mpadge/mapscanner/blob/master/CONTRIBUTING.md).

> Moreover, in some places there is a space and in others there isn't.

I think I have found and rectified all previous non-spaced instances.

> Similarly, if() functions should include {} if they extend to the next line

That's just another personal preference of mine to exclude `{}` for single-line clauses, but I do ensure that these have free lines above and below to ensure ease of reading. This reduces total code length by 1 line for each of my single-line `if` clauses.

> This is more of a matter of preference, but for functions from packages that are used repeatedly throughout a function, I prefer @importFrom rather than ::. The code is cleaner and easier to read. I would do this with sf and dplyr and possibly purrr.

Sorry @potterzot, but I have to disagree on that one, for reasons that have been very well established in [C++](https://stackoverflow.com/questions/1452721/why-is-using-namespace-std-considered-bad-practice). In short, using `@importFrom` generates potential namespace confusion, while `::` remains the only way in R to definitively avoid that.

## Code

> this "tell Mike" statement should be expanded into an explicit message or removed entirely.

Done, thanks.

> Reflecting this comment, handle setting of API key in an informative way.

Done as described at length above.

## Line extraction

Quotes from @potterzot:

> While the package is explicitly aieed at researchers, another use case would be hikers and runners who want to plan a route but don't want to click on an online map. This would let you draw the route, scan it in to the map, and have it uploaded as a shape. If it could be then converted to a GPX file that would be really neat.

followed by comment:

> Adding line extraction would be really neat, but I also don't think it needs to be done as part of the review process. Maybe a later feature down the road.

I will almost certainly defer this until the end of the review process, as extracting continuous lines from raster data is, as I indicated in my previous comment, exceedingly difficult.
