### 2019-09-13 mpadge/mapscanner#17 @khondula
error converting to spatial format
re:  https://github.com/ropensci/software-review/issues/330#issuecomment-528962487
