### 2019-09-21 ropensci/mapscanner#17 @mpadge
Thanks @khondula. This seems to be due to lines being drawn or scanned in very thin, so that the raster representations actually become disconnected. The (default) underlying routines currently presume all features are entirely contiguous in the raster representation, and so fail on these. Of course, for the package to be usable, they should not fail at all, so I need to fix this. I'm working of ways of stitching fragmented pieces back together ...
