### 2019-09-23 0878fac
add user interrupt to get_components
