### 2019-09-24 fa88ade
add ms_rotate_map fn
