+++ _pkgdown.yml
@@ -15,6 +15,8 @@ reference:
   - title: Bonus Features
     contents:
         - ms_aggregate_polys
+        - ms_rotate_map
+        - set_mapbox_token
 
 articles:
   - title: The mapscanner package
