### 2019-09-24 57be483
:wrench:warning: unused typedef 'circ_elem_type'
