### 2019-09-25 0a9dc4c
complete rewrite of extract_channel
now uses functions to automatically estimate number of components and threshold
