+++ R/rectify-maps.R
@@ -100,19 +100,13 @@ ms_rectify_maps <- function (map_original, map_modified, non_linear = 1,
     }
     nr <- m_niftyreg (map_scanned, map, non_linear = non_linear)
     if (!quiet)
-    {
-        message ("\r", cli::symbol$tick, " rectifying the two maps ")
+        message ("\r", cli::symbol$tick, " rectified the two maps  ")
 
-        message (cli::symbol$pointer, " extracting drawn objects ",
-                 appendLF = FALSE)
-    }
-    img <- extract_channel (nr)
+    img <- extract_channel (nr, quiet = quiet)
+
     if (!quiet)
-    {
-        message ("\r", cli::symbol$tick, " extracting drawn objects ")
         message (cli::symbol$pointer, " converting to spatial format ",
                  appendLF = FALSE)
-    }
     res <- rectify_channel (img, f_orig, type = type, n = downsample,
                             concavity = concavity,
                             length_threshold = length_threshold)
@@ -165,33 +159,26 @@ m_niftyreg <- memoise::memoise (function (map_scanned, map, non_linear)
 # threshold is based on delta, which is the aggregate difference between colour
 # channels. Values above (threhold * median (delta)) are considered to be a
 # single colour and are kept for futher analysis; all other values are removed.
-extract_channel <- function (nr, nitems = NULL)
+extract_channel <- function (nr, nitems = NULL, quiet = FALSE)
 {
     img <- nr$image # house and img have 3 layers [r, g, b]
-    threshold <- get_channel_diff_threshold (img)
-    cmat <- get_component_mat (img, threshold)
+    x <- get_channel_diff_threshold (img, quiet = quiet)
+    if (is.null (nitems))
+        nitems <- x$ncomps
+    cmat <- get_component_mat (img, x$threshold)
     tc <- table (cmat [cmat > 0])
-    if (any (tc < 10) & length (tc) > 4) #  > 4 components
-    {
-        tc2 <- table (tc)
-        # names (tc2) gives the sizes of different clusters in ascending order.
-        # The following code finds the first value at which these sizes make a
-        # positive jump, and takes that as the *smallest* size of clusters to be
-        # analysed.
-        x <- diff (as.integer (names (tc2))) # diff between consecutive sizes
-        d3 <- diff (diff (x)) # curvatures of those differences
-        # n is location in names (tc2) of first point of positive curvature in
-        # size
-        n <- as.integer (names (tc2 [which (d3 > 0) [1] + 3]))
-        comps <- names (tc [which (tc >= n)])
-        small_comps <- names (tc) [which (!names (tc) %in% comps)]
-        cmat [which (cmat %in% as.integer (small_comps))] <- 0
-    }
-    cmat [which (cmat > 0)] <- 1
+    comp_nums <- as.integer (names (sort (tc, decreasing = TRUE)) [1:nitems])
+    cmat [which (!cmat %in% comp_nums)] <- 0
+    cmat [which (cmat > 0)] <- 1 # same as %in% comp_nums, but more efficient
+
+    cmat <- t (apply (cmat, 2, rev))
 
     return (cmat)
 }
 
+# threshold is for aggregate difference between the 3 channels, used to
+# distinguish coloured from grey-scale pixels. Higher values extract fewer
+# pixels.
 get_component_mat <- function (img, threshold)
 {
     res <- array (0, dim = c (dim (img) [1], dim (img) [2]))
@@ -205,21 +192,95 @@ get_component_mat <- function (img, threshold)
     rcpp_components (res)
 }
 
-get_channel_diff_threshold <- function (img, nitems = NULL, threshold = 20)
+# terminology note: user interface refers to "items", but code itself identifies
[…]
