### 2019-09-25 3ac323a
check rotation of image files in ms_rectify_maps
