### 2019-09-25 397fe9b
auto-reduce size of image files
