### 2019-09-25 8041d8f
expose nitems param in rectify_maps
