### 2019-09-25 1ac3d29
improve threshold detection; catch fail cases
