### 2019-09-25 #330 @mpadge
So @khondula very kindly sent me all of the maps mentioned above. All of them were complete breaking cases, and addressing them has helped enormously to make the package much more robust. Lots of the internal mechanisms have been completely reconceived. The package now makes an initial guess at number of items to be extracted, then estimates a signal-to-noise threshold to optimally identify those features. There is also an additional optional parameter `nitems` which, if specified, overrides the automatic estimation, and can improve results for difficult cases. The following table summarises what now happens for each of the previously failing files:

file                            | result?
------------------------------- | -------
choptank-annotated.pdf          | ok
choptank-annotated.png          | ok
choptank-annotate2.png          | ok
choptank-annotate3.pdf          | ok
choptank-annotate4.pdf          | ok
choptank-annotate-5.pdf         | auto-detects only 5 items and kinda fails on one, but works perfectly with `nitems = 6`
choptank-annotate-line.png      | works perfectly as polygon (but actual line-detection not yet implemented; see #5)
choptank-annotate-points.png    | works perfectly as points, hulls, and polygons
omaha-pink2.pdf                 | auto-detects only 1 item but then errors appropriately; specifying `nitems = 2` works
omaha-pink.pdf                  | works, but inaccurately because fails to trim border

@khondula could you please confirm that all is okay? And if so, it'd be nice if you could close the [corresponding issue on the package repo](https://github.com/mpadge/mapscanner/issues/17) itself. Thank you so much for helping to greatly improve the package already!
