### 2019-09-25 98951a9
remove boundary polygons with < 4 points
