### 2019-10-07 9c47bc0
default rectify type = hulls, not polygons
