### 2019-10-10 #330 @khondula
Hi @mpadge, your updates seem to have worked! All cases were able to convert to spatial format, and the messages regarding `nitems` and `ms_rotate_map` were very helpful. I will close issue on package repo.

One new thing I noticed however is that with `type = "hulls"` (not what I intended but discovered because it's the default!), the output objects seem to be rotated 90 degrees - I think this might be happening [here](https://github.com/mpadge/mapscanner/blob/9c47bc0c44eff351135f0ac6425639be0dea81e1/R/rectify-maps.R#L342)?

Using the images from the example as f_orig and f_mod:

```
res_hulls <- ms_rectify_maps(f_orig, f_mod, type = "hulls", nitems = 2)
res_polygons <- ms_rectify_maps(f_orig, f_mod, type = "polygons", nitems = 2)
```

res_hulls is rotated:
```
res_hulls %>%
  leaflet() %>%
  addTiles() %>%
  addPolygons()
```

res_polygons is fine though
```
res_polygons %>%
  leaflet() %>%
  addTiles() %>%
  addPolygons()
````
