### 2019-10-10 ropensci/mapscanner#17 @khondula
closing as per [review comment](https://github.com/ropensci/software-review/issues/330#issuecomment-540320930) confirming it works!
