### 2019-10-10 #330 @mpadge
Thanks so much for checking @khondula - that's great news that it all works. And sorry about the rotation - it's something I only just discovered and am working on right now. I'd spent so much time on `type = "polygons"` that I had neglected to ensure `"hulls"` were also appropriately rotated. Should be fixed very soon ... I'll give a full response to your review asap
