### 2019-10-10 mpadge/mapscanner#21 @mpadge
hulls are rotated the wrong way
As indicated by @khondula in [review comment](https://github.com/ropensci/software-review/issues/330#issuecomment-540320930). `type = "polygons"` works, but `type = "hulls"` rotates the result incorrectly. Fix
