### 2019-10-10 e07335a
fix hull rotation; closes #21
