### 2019-10-12 mpadge/mapscanner#22 @mpadge
update docs to indicate it works with digial maps
As indicated in review comments of @khondula in [rOpenSci review](https://github.com/ropensci/software-review/issues/330#issuecomment-528962487), the package actually works perfectly well with maps in digital form, yet all docs refer to printing and hand-drawing on maps. Update and explicitly describe process of digital annotation and rectification.
