### 2019-10-21 01b6ed9
message on successful set_mapbox_token
