### 2019-10-21 mpadge/mapscanner#23 @mpadge
fix api key handling
In response to error messages issued during [review ](https://github.com/ropensci/software-review/issues/330#issuecomment-528962487) by @khondula:
```
Error in get_mapbox_token() :
  Map generation requires a mapbox API key to be set with Sys.setenv with a name that includes either the strings 'mapbox' or 'mapscanner'
In addition: Warning message:
In if (length(tok) != 1 | tok == "") stop("Map generation requires a mapbox API key to be set with ",  :
  the condition has length > 1 and only the first element will be used
```
