### 2019-10-21 5858b5e
fix token handling; closes #23
