+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: mapscanner
 Title: Print Maps, Draw on Them, Scan Them Back In
-Version: 0.0.1.002
+Version: 0.0.1.003
 Authors@R: c(
     person("Mark", "Padgham", email="mark.padgham@email.com", role=c("aut", "cre")),
     person("Michael D.", "Sumner", role = "aut", email = "mdsumner@gmail.com"),
