### 2019-10-21 5ea5254
rm system.time wrapper from example [skip ci]
