+++ R/aggregate_polys.R
@@ -24,7 +24,7 @@
 #' set.seed(6)
 #' pts <- expand.grid (x = 1:8, y = 1:10) %>% st_as_sf (coords = c("x", "y"))
 #' xsf <- sf::st_buffer (pts, runif (nrow (pts), 0.2, 1.5))
-#' # system.time (out <- ms_aggregate_polys (xsf))
+#' # out <- ms_aggregate_polys (xsf)
 ms_aggregate_polys <- function (px, ...)
 {
     tri_map <- triangulate_map_sf (px)
