### 2019-10-21 4ccd53a
do not check token if rasterBrick provided
