+++ R/generate-maps.R
@@ -43,13 +43,15 @@ ms_generate_map <- function (bbox, max_tiles = 16L, mapname = NULL,
 
     style <- match.arg (tolower (style), c ("light", "streets", "outdoors"))
 
-    # used here just to confirm token exists:
-    mapbox_token <- get_mapbox_token ()
 
     if (is.null (raster_brick))
+    {
+        # used here just to confirm token exists:
+        mapbox_token <- get_mapbox_token ()
         raster_brick <- get_raster_brick (bbox = bbox,              # nocov
                                           max_tiles = max_tiles,    # nocov
                                           style = style)            # nocov
+    }
 
     fname_p <- map_to_pdf (raster_brick, mapname)
     fname_j <- map_to_png (raster_brick, mapname)
