### 2019-10-21 4805f3a
rm reference to rectification requiring png from readme; as per review [skip ci]
