### 2019-10-21 e2b159b
add explicit mapview/deck example code to readme
