+++ src/components.cpp
@@ -78,7 +78,7 @@ void getNeighbours (
                j_k = static_cast <size_t> (j + dy [k]);
         if (i_k < 0 || j_k < 0 || i_k > (image.nrow () - 1) ||
                 j_k > (image.ncol () - 1))
-            continue;
+            continue; // # nocov
         if (image (i_k, j_k) && comp_mat (i_k, j_k) == 0)
         {
             std::pair <int, int> nbPair = std::make_pair (i + dx [k], j + dy [k]);
