### 2019-10-22 7de40b5
add test-file-manip
