### 2019-10-22 42a1dfb
tweak readme to indicate possibility of screen-based usage
