### 2019-10-22 49380cc
mention possibility of screen-based usage in vignette
