### 2019-10-22 d8ac7cc
extend description of type = polygons
