### 2019-10-22 #330 @mpadge
## Response to review by @khondula

Mostly in the form of quotations of review comments, followed by my own comments and descriptions of responses and associated improvements to the package.

### General

> This is a great package with many potential uses and I'm excited to see it being developed!

Thanks for the encouraging start - I'm excited about having developed it, and the ongoing improvements brought about by this review process.

> Unfortunately I was not yet able to get the main functionality working using a scanned in document, as every attempt seems to stall out at the "converting to spatial format" step (details below). I look forward to guidance from the authors to help diagnose my issues and get it working. Assuming these can be resolved ...

Hopefully all resolved via the above comments in this review, and in [issue#17 in the package repo](https://github.com/mpadge/mapscanner/issues/17).

> ... my only other concerns are some areas where the documentation could be improved, and a question about whether there is a need for a mapbox account.

Responded to in detail below.

### Specific comments

> Would it be possible to return a message if 'set mapbox token' is completed successfully?

Messages on success or fail of `set_mapbox_token()` added in [this commit](https://github.com/mpadge/mapscanner/commit/01b6ed9be0a4bb4332ad8bbc6aaf38d869ed14ea).

> I also prefer not having a space between function names and opening parentheses, but defer to author preference.

Respectfully noted, but i like white space. The style is noted in [`CONTRIBUTING.md`](https://github.com/mpadge/mapscanner/blob/master/CONTRIBUTING.md), and exists simply because it is easier to read
```r
this <- function1 (function2 (x))
```
than
```r
this <- function1(function2(x))
```

> It was unclear to me what the MapBox token is used for. If it is only necessary for acquiring tiles as a rasterbrick, would it be possible to use the rosm package as below, which I don't believe requires signing up for any external account?

This would be a good idea but for two issues which I believe prevent it being able to be implemented:

1. The `rosm` package ['will probably be retired in the next year']( https://github.com/paleolimbot/rosm/) (comment added on [Jan 16, 2019](https://github.com/paleolimbot/rosm/commit/0607a5d7c5f4c1b36fef11af893162aae279be2a#diff-04c6e90faac2675aa89e2176d2eec7d8R37)).
2. That package avoids the use of an API key through using the third-party service [`a.basemaps.cartocdn.com`](http://a.basemaps.cartocdn.com), which offers only a restricted set of three map styles, none of which are particularly suitable for creating black-and-white maps needed by this package.

Of course it would be a great convenience boost not to require an API key, but at present I see no alternative alas.

> I had trouble set the mapbox key as an environment variable in my ~/.Renviron file, error copied below. I did not have any problems using the set_mapbox_token method.

```{r}
Error in get_mapbox_token() :
  Map generation requires a mapbox API key to be set with Sys.setenv with a name that includes either the strings 'mapbox' or 'mapscanner'
In addition: Warning message:
In if (length(tok) != 1 | tok == "") stop("Map generation requires a mapbox API key to be set with ",  :
  the condition has length > 1 and only the first element will be used
```

This was shoddy programming on my part, and has been fixed as detailed in [issue #23](https://github.com/mpadge/mapscanner/issues/23).

> When the modified/annotated map includes polygons but the type argument is "points", would it be possible to include more helpful error message, such as a suggestion to "try type = hull"?

```{r}
> res <- ms_rectify_maps(f_orig, f_mod, type = "points")
══ mapscanner ══════════════════════════════════════════════════════════════════════════
✔ rectifying the two maps
✔ extracting drawn objects
❯ converting to spatial format Error in CPL_geos_op("centroid", x, numeric(0), integer(0), numeric(0),  :
  Evaluation error: IllegalArgumentException: point array must contain 0 or >1 elements.
```

I suspect this would have reflected the generally processing workflow that previously failed on most of @khondula's tests. The work undertaken to ensure that these all work should also have ensured that errors like this do not arise.

## Documentation

From readme:

> "generally be the most convenient for printing, while the rectification itself requires .png-format images" - My understanding is that the conversion to png happens anyway if needed, so this could be taken out of the general intro and only left in the more detailed info.

That comment removed in [this commit](https://github.com/mpadge/mapscanner/commit/4805f3a8998d1118e0029fb3ab538207a7686552)

> For the initial example, is it potentially confusing to have the main function call wrapped in system.time, since that is not necessary for running the function?

Thanks for finding that - the `system.time` call has been removed in [this commit](https://github.com/mpadge/mapscanner/commit/5ea5254bf93418af59b6b826f17de2fad1376abd).

> "Finally, we can plot the result as an interactive map using packages like mapdeck, or mapview:" Would it be possible to show an example of this code for users unfamiliar with those packages?

Good idea - code added in [this commit](https://github.com/mpadge/mapscanner/commit/e2b159b581b6d01edadb06a65b217aecd1edd8e3).

> I agree with the statements about the use of paper maps, however wouldn't there also be use cases where digital annotations (e.g. with a stylus, finger, pdf editor) on a phone/tablet/other device could be saved and used with this package? If so, perhaps some language could be included to clarify that is possible as well.

That is of course very important. I had been so busy conceiving of the package as a tool to aid drawing on actual paper that i hadn't even sufficiently considered how direct the transfer is to screen devices. The text of the `README` has accordingly been updated in [this commit](https://github.com/mpadge/mapscanner/commit/42a1dfb362f02a47e730d79dc2236a72badaae6e), with corresponding changes in the vignette made in [this commit](https://github.com/mpadge/mapscanner/commit/49380ccdf394167770d06ae5a25ca3b7454e8aa9).

> It is unclear what is meant by a "practical workflow". The current language under 'practical tips' seems to imply that other software packages are meant to not be practical. Would it be possible to either clarify, or leave out the phrase about 'unlike most software packages...'? I appreciate that there is a specific workflow/use case in mind, but it is also clear that there are many more potential applications!

Yes, a very good point. I have simply removed that phrase, so that section now starts, "The `mapscanner` package is intended to aid a *practical* workflow, and so a few practical tips may be recommended here ..."

From vignette:

> "enables maps to be printed out" - The wording could be more precise here. I believe it would it be more accurate to say something along the lines of "enables printed out maps to be drawn on and..." since the functions aren't actually doing the printing.

Changed to, "enables lines drawn by hand on maps to be converted to spatial objects," with references to printing only hypothetical for paper-based use.

> Is it possible to add more details to clarify the distinction between type = hulls vs type = polygons?

Description extended within the vignette in [this commit](https://github.com/mpadge/mapscanner/commit/d8ac7ccb758f930917701c21ad26fcc3dc3c9e29).

### Other notes

> One warning during local installation:

```{r}
In file included from concaveman.cpp:1:
./concaveman.h:587:38: warning: unused typedef 'circ_elem_type' [-Wunused-local-typedef]
    typedef CircularElement<Node<T>> circ_elem_type;
                                     ^
1 warning generated.
```

That warning is because of C++14 interpretation of a C++11-standard usage, and depends on your local compiler, in particular any flags you may have set -- or indeed not set -- in `~.R/Makevars`. It's a strictly local issue, not one associated with the package itself.

> Warning from github install:

```
Warning: invalid uid value replaced by that for user 'nobody'
Warning: invalid gid value replaced by that for user 'nobody'
```

This issue is described in [this R-devel mailing list post](https://r.789695.n4.nabble.com/tar-warnings-in-R-3-0-2-RC-when-R-is-installed-by-a-different-non-root-user-td4676765.html), and [this StackOverflow Q](https://stackoverflow.com/questions/30599326/warning-message-during-building-an-r-package-invalid-uid-value-replaced-by-that). It can safely be ignored.

> Error in devtools::check()

```
Loading mapscanner
Error in getDLLRegisteredRoutines.DLLInfo(dll, addNames = FALSE) :
  must specify DLL via a “DLLInfo” object. See getLoadedDLLs()
```

I guess you're checking on windows? If that problem persists, the usual approach would be to confirm that everything is okay with
```bash
R CMD build mapscanner
R CMD check mapscanner_0.X.X.tar.gz
```
I don't have access to a windows machine to be able to replicate this. @sckott do you perchance have access to such? And if so, would you perhaps be able to run a quick `devtools::check()` on the repo to confirm or otherwise whether this persists?

----

### concluding comments

Thank you very much @khondula for a very constructive review, and in particular thank you for privately sending me all of your test maps which previously failed, and which led to me finding quite a number of holes and glitches in former code. Fixing all of those has been the single greatest improvement to package functionality since its inception. I remain hugely indebted, and extremely grateful!
