### 2019-10-22 d3648b7
screen dump final count of objects; closes #24
