### 2019-10-22 f7e39a8
add internal doc (https://github.com/ropensci/software-review/issues/330#issuecomment-528304975), and minor fix
