### 2019-10-23 #330 @mpadge
Thanks @potterzot. I am unfortunately unable to reproduce that failing test - could you maybe try re-installing and ensure that you get
``` r
devtools::load_all (".", export_all = FALSE)
#> Loading mapscanner
packageVersion("mapscanner")
#> [1] '0.0.1.5'
```

All tests current pass on my local machine, on [travis](https://travis-ci.org/mpadge/mapscanner), and on [appveyor](https://ci.appveyor.com/project/mpadge/mapscanner).
