### 2019-10-23 #330 @khondula
All looks good to me - changes work successfully and the `devtools::check()` error has gone away. Thanks for this package @mpadge !
