### 2019-10-23 #330 @potterzot
I've updated all of my R packages, started with a fresh git clone, and am still getting that error and am unable to build the package. I'm not sure why, but given that I'm the only one, this seems like an error on my part. Documenting those issues here in case there is some build difference that is relevant. Or if there's something I'm doing wrong please let me know!

I've updated my review to suggest accepting, great improvements and good package @mpadge!

One suggestion that I have after trying to debug:

- In `tests/testthat/test-file-manip.R` you check whether the png files are created by taking a difference in files found in `f1` and `f2`. If the test is run interactively multiple times, `f1` ends up including the `junk.png` and `test.png` files that are created as part of the test. I would suggest either just testing directly for those files or explicitly removing the files afterward as part of the test.

Running `R -e 'devtools::test()' produces these errors:

```{r}
...
test-file-manip.R:14: failure: file manipulation
`ms_rotate_map(f_orig, f_modified)` produced messages.
...
test-rectify.R:50: error: rectify
'from' must be a finite number
1: expect_silent(res_p <- ms_rectify_maps(f_orig2, f_modified2, type = "polygons", quiet = TRUE)) at /home/potterzot/reason/projects/ropensci/mapscanner/tests/testthat/test-rectify.R:50
2: quasi_capture(enquo(object), NULL, evaluate_promise)
3: .capture(act$val <- eval_bare(get_expr(.quo), get_env(.quo)), ...)
4: withr::with_output_sink(temp, withCallingHandlers(withVisible(code), warning = handle_warning,
       message = handle_message))
5: force(code)
6: withCallingHandlers(withVisible(code), warning = handle_warning, message = handle_message)
7: withVisible(code)
8: eval_bare(get_expr(.quo), get_env(.quo))
9: ms_rectify_maps(f_orig2, f_modified2, type = "polygons", quiet = TRUE)
10: rectify_channel(img, f_orig, type = type, n = downsample, concavity = concavity,
       length_threshold = length_threshold, quiet = quiet) at /home/potterzot/reason/projects/ropensci/mapscanner/R/rectify-maps.R:121
11: polygon_boundaries(channel) at /home/potterzot/reason/projects/ropensci/mapscanner/R/rectify-maps.R:358
12: get_shape_index(cmat_i, x = TRUE) at /home/potterzot/reason/projects/ropensci/mapscanner/R/rectify-maps.R:409
13: seq(min(index) - 1, max(index) + 1) at /home/potterzot/reason/projects/ropensci/mapscanner/R/rectify-maps.R:428
14: seq.default(min(index) - 1, max(index) + 1)
15: stop("'from' must be a finite number")
```

Running `R CMD build .` or `R -e 'devtools::build()' produces the following error:

```{r}
--- re-building ‘mapscanner.Rmd’ using rmarkdown
   Loading required namespace: sf
   File mapshot-polys.png not found in resource path
   Error: processing vignette 'mapscanner.Rmd' failed with diagnostics:
   pandoc document conversion failed with error 99
   --- failed re-building ‘mapscanner.Rmd’
```
