### 2019-10-24 d9389e1
add missing vignette file for ropensci/software-review#330
