### 2019-10-24 #330 @mpadge
@potterzot I'm so sorry about your continued troubles there. The commit referenced immediately above solves your latter problem of not being able to install. About your other issue from `devtools::test()`, i can not unfortunately do much. I have managed to generate (very) occasional problematic behaviour using this command, which gives in my case the very uninformative message,
```
#> Error in if (length(xx) > 0) { : argument is of length zero
```
This error is not issued by `mapscanner` (`grep -r "xx" .` returns nothing - there is no `"xx"` in the code), so presumably arises from the depths of `devtools::test()` itself. Do all tests run okay for you when you run `devtools::check()`? Or `R CMD check`?
