### 2019-10-26 #330 @potterzot
I spent some time debugging. The tests in `test-rectify.R` fail because they lack a "nitems = 2" parameter call and the internal code finds the number of components to be zero. I think that this is due to [lines 284-287 of `R/rectify-maps.R`](https://github.com/mpadge/mapscanner/blob/d9389e1b78fa1bc3be8e7085e90492e56fb627dd/R/rectify-maps.R#L284-L287) not allowing for the case in which all values of `dn` are __less than or equal to__ 0, and as a result `thr0` ends up as NA if `dn` has some 0 values and some negative values. Anyway, if I change line 284 to be `<=`, the tests pass. I've submitted a PR.

As to why this works fine on travis but was failing on my laptop I can't say. The only thing I could think of was perhaps cpp compilers are different but that seems like an obscure error. My `Makevars` is completely empty for building this, so I shouldn't have any weird tweaks.

### A second issue

In the process I realized the vignette contains a lot of hardcoded outputs. I would suggest making these outputs the actual output from the code block. This bit me because when trying to build the vignette I thought it was reporting the actual output of the code block but it wasn't.
