### 2019-10-28 #330 @sckott
Approved! Thanks @mpadge for submitting and @khondula and @potterzot for your reviews!

A few comments:

- link to the COC from the README
- on check, there's a package size warning on my macos that it's 5.2 mb, maybe compressing some images may bring that below 5?
- i couldn't try the mapdeck eg because couldn't get the dev versions of it's dependencies to install from source to get required versions - no worries, let's move on

- [ ] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo. you should be able to transfer, let me know if it doesn't work.
"```
- [ ] If you already have a `pkgdown` website, fix its URL to point to `https://docs.ropensci.org/package_name` and deactivate the automatic deployment you might have set up, since it will not be built centrally like for all rOpenSci packages, see http://devdevguide.netlify.com/#docsropensci. In addition, in your DESCRIPTION file, include the docs link in the `URL` field alongside the link to the GitHub repository, e.g.: `URL: https://docs.ropensci.org/foobar (website) https://github.com/ropensci/foobar`
- [ ] Add a mention of the review in `DESCRIPTION` via [`rodev::add_ro_desc()`](https://docs.ropensci.org/rodev/reference/add_ro_desc.html).

For JOSS:
- [ ] Submit to JOSS at <https://joss.theoj.org/papers/new>, using the rOpenSci GitHub repo URL. When a JOSS "PRE REVIEW" issue is generated for your paper, add the comment: `This package has been reviewed by rOpenSci: https://LINK.TO/THE/REVIEW/ISSUE, @ropensci/editors`
