### 2019-10-28 mpadge/mapscanner#27 @mpadge
reduce package installed size
Currently sometimes exceeds 5MB, as noted by @sckott in [review](https://github.com/ropensci/software-review/issues/330#issuecomment-547041251), but this is due to `ggplot2` graphics in the main vignette:
``` r
du <- function (where) {
    x <- system (paste ("du -sh ", where), intern = TRUE)
    message (paste (x, collapse = "\n"))
}
du ("*") # in package repo
#> 928K inst
#> 500K man
#> 1.4M R # mostly local file histories and stuff
#> 1.5M src # same
#> 268K tests
#> 600K vignettes
setwd ("<path>/<to>/<installed>/<package>")
du ("*")
#> 4.3M inst
#> 500K man
#> 72K  R
#> 56K  src
#> 92K  tests
#> 2.0M vignettes
du ("inst/*")
#> 3.4M inst/doc
#> 900K inst/extdata
#> 4.0K inst/paper.md
du ("inst/doc/*")
#> 3.4M inst/doc/mapscanner.html
#> 8.0K inst/doc/mapscanner.R
#> 24K  inst/doc/mapscanner.Rmd
```

And about half of the installed package size (3.4/7 MB) is the *rendered* version of the vignette. The original (repo) directory is only 600K, 380K of which is `vignettes/omaha-polygons.png`. The `ggplot2` images demonstrating [polygon aggregation](https://github.com/mpadge/mapscanner/blob/master/vignettes/mapscanner.Rmd#L375-L393) inflate the package to this unwieldy size. Fiddling with `fig.width` and `fig.height` options in Rmarkdown chunks does nothing either - the size is inherent in the `.eps` info contained within the image itself. Why so big? And what to do about it? ... resort to base graphics?
