### 2019-10-28 #330 @mpadge
Thanks @sckott, that is indeed fabulous news. Before proceeding any further, however, I've opened an [issue in the repo](https://github.com/mpadge/mapscanner/issues/27) on something that's been bugging me, and which i want to solve before finishing up here - and which you yourself raised:

> on check, there's a package size warning on my macos that it's 5.2 mb, maybe compressing some images may bring that below 5?

It's not the images, it's `ggplot2`. In particular, the [three figures in the vignette](https://mpadge.github.io/mapscanner/articles/mapscanner.html#bonus-feature-polygon-aggregation) which illustrate the peripheral yet (in our opinions) rather cool feature of polygon aggregation. `ggplot2` simply results in huge figure sizes here, and it seems nothing can be done about that, other than to maybe switch to base graphics. But a solution is needed, because there is already another vignette demonstrating the different mapstyles available, and the most important one in the pipeline to arise from a large sociological study currently being conducted with the package ... but neither of those vignettes can actually be included with the package until we can reduce the package size, which it seems can't be done with `ggplot2` being used. Any and all assistance greatly appreciated, even if only advice to ditch `gg` and revert to good old base.
