### 2019-10-28 #330 @potterzot
@mpadge as far as the ggplot size issue goes, one option may be to use the `dpi`, `height`, and `width` parameters of `ggsave()` to save the images and then read them in, rather than displaying them directly. I think this would give you control over the file size, but still allow dynamic generation of the figures.
