### 2019-10-28 ropensci/mapscanner#27 @mdsumner
drop the png in as a static bitmap and turn off eval for the ggplot?
