### 2019-10-28 ropensci/mapscanner#27 @mpadge
oh yeah, that should do it. Thanks :rocket:
