### 2019-10-29 ropensci/mapscanner#27 @mpadge
@mdsumner this one's for future reference: it actually comes from `mapview::mapshot`, which itself just calls `webshot::webshot` and uses default options which are `vwidth = 922` and `vheight = 744` used to determine sizes of viewport to be captured. That results in really large image sizes. Luckily, @tim-salabim foresaw the need to pass additional parameters through to `webshot`, so `mapshot` allows these to be adjusted/reduced. That solves it. Phew.
