### 2019-10-29 5e6e9dc
reduce mapshot image sizes for #27
