### 2019-12-01 c2a139b
update vignette
