+++ vignettes/mapscanner.Rmd
@@ -16,6 +16,7 @@ vignette: >
 
 ```{r pkg-load, echo = FALSE, message = FALSE}
 library (mapscanner)
+library (ggplot2)
 requireNamespace ("sf")
 # vwidth/height are passed via mapview::mapshot to webshot::webshot,
 # and default to (992,744), which bloats package size (issue #27).
@@ -23,6 +24,24 @@ red <- 2
 vwidth <- 992 / red
 vheight <- 744 / red
 ```
+
+<!--
+A NOTE regarding the size of this vignette, for which it was difficult to
+produce a sufficiently small sized end result, largely because of the mapshot
+images. The above values can be passed to mapshot to reduce those. The
+following table shows sizes of resultant installed versions following various
+options:
+
+ggplot | mapshot } inst size | vignettes size
+---  | --- | --- | ---
+ -   |  -  | 2.2MB | 1.2MB
+ yes |  -  | 2.4MB | 1.5MB
+ yes |  1  | 2.9MB | 2.0MB (original mapshot image only)
+ yes |  3  | 3.4MB | 2.5MB
+
+ At present, the vignette is rendered as in the final row, but can potentially
+ be reduced in size in the future according to this table. -->
+
 
 ## What does this package do for me?
 
@@ -232,14 +251,14 @@ rectification, while the second has two red lines drawn upon it:
 <!--
 <img src="omaha-polygons.png" width="600" height="400"/>
 -->
-```{r omaha-poly-png, echo = FALSE, out.width = "75%"}
+```{r omaha-poly-png, echo = FALSE, out.width = "75%", eval = TRUE}
 knitr::include_graphics ("omaha-polygons.png")
 ```
 
 
 Converting the lines on this scanned image file is then as simple as:
 
-```{r rectify}
+```{r rectify, message = FALSE}
 f_orig <- system.file ("extdata", "omaha.png", package = "mapscanner")
 f_mod <- system.file ("extdata", "omaha-polygons.png", package = "mapscanner")
 xy <- ms_rectify_maps (f_orig, f_mod, nitems = 2)
@@ -269,7 +288,9 @@ mapview::mapview (xy)
 -->
 ```{r mapshot-hulls-png, echo = FALSE, out.width = "75%"}
 knitr::include_graphics ("mapshot-hulls.png")
+#plot (xy)
 ```
+
 
 
 ### Types of map markings and types of spatial objects
@@ -334,7 +355,9 @@ mapview::mapview (xy)
 -->
 ```{r mapshot-polys-png, echo = FALSE, out.width = "75%"}
 knitr::include_graphics ("mapshot-polys.png")
+#plot (xy)
 ```
+
 
 
 ```{r type-points}
@@ -362,7 +385,9 @@ mapview::mapview (xy)
 -->
 ```{r mapshot-points-png, echo = FALSE, out.width = "75%"}
 knitr::include_graphics ("mapshot-points.png")
+#plot (xy)
 ```
+
 
 
 
@@ -403,7 +428,7 @@ polys <- rbind (polys, aggr)
 library (ggplot2)
 ggplot (polys, aes (fill = n)) + geom_sf () + facet_wrap (~type)
 ```
-```{r random-polys, echo = FALSE}
+```{r random-polys, echo = FALSE, eval = TRUE}
 n <- 5 # number of polygons
 polys <- lapply (seq (n), function (i) {
                      nxy <- 20 # number of points used to generate hull
@@ -416,7 +441,6 @@ aggr <- ms_aggregate_polys (polys)
 polys$type <- "raw polygons"
 aggr$type <- "aggregated polygons"
 polys <- rbind (polys, aggr)
-library (ggplot2)
 theme_set (theme_minimal ())
 labeller <- c (
[…]
