### 2019-12-01 5521391
update man entries to roxygen 7
