+++ man/ms_rectify_maps.Rd
@@ -4,9 +4,17 @@
 \alias{ms_rectify_maps}
 \title{ms_rectify_maps}
 \usage{
-ms_rectify_maps(map_original, map_modified, nitems = NULL,
-  non_linear = 1, type = "hulls", downsample = 10, concavity = 0,
-  length_threshold = 10, quiet = FALSE)
+ms_rectify_maps(
+  map_original,
+  map_modified,
+  nitems = NULL,
+  non_linear = 1,
+  type = "hulls",
+  downsample = 10,
+  concavity = 0,
+  length_threshold = 10,
+  quiet = FALSE
+)
 }
 \arguments{
 \item{map_original}{File name of the original map without anything drawn over
