### 2020-01-24 #330 @mpadge
Yeah, I still have to do the `paper.md` for JOSS. I've been waiting for some empirical results from a currently-running study, but that's not forthcoming, so I'll write an alternative one instead. I'll let you know as soon as that's ready, and then I think we'll be done. Hopefully in just a few days time.
