+++ paper.md
@@ -1,0 +1,32 @@
+---
+title: mapscanner
+tags:
+    - maps
+    - spatial
+    - R
+    - Simple Features
+authors:
+    - name: Mark Padgham
+      affiliation: 1
+affiliations:
+    - name: Active Transport Futures, Muenster, Germany
+      index: 1
+date: 26 January 2020
+#bibliography: vignettes/mapscanner-refs.bib
+nocite: |
+  @*
+---
+
+# Summary
+
+`mapscanner` is an **R** package that enables lines drawn by hand on maps to be
+converted to spatial objects. The package has two primary functions: one for
+producing maps, and one for rectifying hand-drawn lines to the coordinate
+system of the original map. The package is intended for use in social surveys
+and similar endeavours in which hand-drawn markings on maps need to be
+converted to spatial objects. Maps can be either paper- or screen-based.
+Markings on paper maps need to be scanned, photographed, or otherwise
+digitised, while maps with screen-based markings need to be saved as
+`.png`-format images.
+
+# References
