+++ README.Rmd
@@ -39,7 +39,7 @@ at [rOpenSci](https://ropensci.org).
 
 `mapscanner` is not (yet) on CRAN. The development version can be installed with
 ```{r install, eval = FALSE}
-remote::install_github("mpadge/mapscanner")
+remotes::install_github("mpadge/mapscanner")
 ```
 The package can then be loaded for usage in a R session with
 ```{r library}
