### 2020-02-02 ropensci/mapscanner#27 @mpadge
Closing now, because they're reduced about as much as can be. Package is still 4.5MB with just the four main map images in the vignette, of which 3.4MB is `inst/`, of which 2.4MB is the main vignette. The 1MB that's not `inst/` is from the 2 images in `man/figures` (mostly the polygon image at 390kB), and from the same image repeated in `vignettes`. The 1MB in `inst` that is not the vignette is also these images copied again in `inst/extdata`. There's really not much that can be done about that, and the only real step to reduce size further would be to make only one of the vignette images actually contain a background map, and all other just containing the various polygon outlines absent background. But that looks really crap compared with the current maps, so happy to roll with things the way they are for the moment.
