### 2020-02-03 63d4745
suggests jpeg, polyclip for #11
