+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: mapscanner
 Title: Print Maps, Draw on Them, Scan Them Back In
-Version: 0.0.1.011
+Version: 0.0.1.012
 Authors@R: c(
     person("Mark", "Padgham", email="mark.padgham@email.com", role=c("aut", "cre")),
     person("Michael D.", "Sumner", role = "aut", email = "mdsumner@gmail.com"),
@@ -18,13 +18,11 @@ Imports:
     curl,
     fs,
     glue,
-    jpeg,
     magrittr,
     magick,
     memoise,
     pdftools,
     png,
-    polyclip,
     purrr,
     raster,
     Rcpp,
@@ -40,11 +38,13 @@ Suggests:
     dplyr,
     gibble,
     ggplot2,
+    jpeg,
     knitr,
     lwgeom,
     mapview,
     mmand,
     osmdata,
+    polyclip,
     rmarkdown,
     testthat,
     spelling
@@ -54,6 +54,6 @@ SystemRequirements:
 VignetteBuilder: knitr
 URL: https://github.com/mpadge/mapscanner
 BugReports: https://github.com/mpadge/mapscanner/issues
-RoxygenNote: 7.0.1
+RoxygenNote: 7.0.2
 Roxygen: list(markdown = TRUE)
 Language: en-GB
