### 2020-02-03 2a450c7
add mapstyles vignette; closes #10
