### 2020-02-03 ropensci/mapscanner#27 @sckott
Sounds good, thanks @mpadge !
