### 2020-02-04 e4ebc68
add mapstyles vignette to _pkgdown
