### 2020-02-06 cc651e3
finish paper.md; closes #28
