### 2020-02-06 #330 @mpadge
Okay @sckott, thanks for your patience, and now all done I think, including just updating the last check item in the initial list to enable [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3647922.svg)](https://doi.org/10.5281/zenodo.3647922). Good to go?
