### 2020-03-03 42cd350
add gitlab,bitbucket,sourcehut repo details
