+++ README.Rmd
@@ -37,8 +37,12 @@ at [rOpenSci](https://ropensci.org).
 
 ## installation
 
-`mapscanner` is not (yet) on CRAN. The development version can be installed with
+`mapscanner` is not (yet) on CRAN. The development version can be installed
+with any of the following options:
 ```{r install, eval = FALSE}
+remotes::install_git("https://git.sr.ht/~mpadge/mapscanner")
+remotes::install_bitbucket("mpadge/mapscanner")
+remotes::install_gitlab("mpadge/mapscanner")
 remotes::install_github("mpadge/mapscanner")
 ```
 The package can then be loaded for usage in a R session with
