+++ README.Rmd
@@ -206,3 +206,9 @@ vignette](https://mpadge.github.io/mapscanner/articles/mapscanner.html) for
 more detail of what the `mapscanner` package can do.
 
 
+## Code of Conduct
+
+Please note that this project is released with a [Contributor Code of
+Conduct](CODE_OF_CONDUCT.md).  By participating in this project you agree to
+abide by its terms.
+
