+++ README.md
@@ -211,3 +211,9 @@ And our hand-drawn lines shown above have been converted to standard
 spatial objects able to be analysed in any desired way. See the [package
 vignette](https://mpadge.github.io/mapscanner/articles/mapscanner.html)
 for more detail of what the `mapscanner` package can do.
+
+## Code of Conduct
+
+Please note that this project is released with a [Contributor Code of
+Conduct](CODE_OF_CONDUCT.md). By participating in this project you agree
+to abide by its terms.
