### 2020-03-19 6b59826
add rOpenSci footer
