+++ README.Rmd
@@ -212,3 +212,5 @@ Please note that this project is released with a [Contributor Code of
 Conduct](CODE_OF_CONDUCT.md).  By participating in this project you agree to
 abide by its terms.
 
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
+
