### 2020-03-19 84c8032
deactivate travis pkgdown deployment
