+++ .travis.yml
@@ -16,9 +16,9 @@ r_packages:
 
 after_success: Rscript -e "covr::codecov()"
 
-before_deploy: Rscript -e 'remotes::install_cran("pkgdown")'
-deploy:
-    provider: script
-    script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
-    skip_cleanup: true
+#before_deploy: Rscript -e 'remotes::install_cran("pkgdown")'
+#deploy:
+#    provider: script
+#    script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
+#    skip_cleanup: true
 
