+++ vignettes/mapscanner.Rmd
@@ -61,9 +61,12 @@ digitised, while maps with screen-based markings need to be saved as
 ### Installation
 
 `mapscanner` is not yet on CRAN, so must be installed with the `remotes`
-package:
+package, via any one of the following options:
 ```{r install, eval = FALSE}
-remotes::install_github ("mpadge/mapscanner")
+remotes::install_github ("ropensci/mapscanner")
+remotes::install_git("https://git.sr.ht/~mpadge/mapscanner")
+remotes::install_bitbucket("mpadge/mapscanner")
+remotes::install_gitlab("mpadge/mapscanner")
 ```
 After that, the library can be loaded with
 ```{r load}
@@ -75,7 +78,7 @@ library (mapscanner)
 The package is designed to enable the following workflow:
 
 1. Generate a map with the
-   [`ms_generate_map()`](https://mpadge.github.io/mapscanner/reference/ms_generate_map.html)
+   [`ms_generate_map()`](https://docs.ropensci.org/mapscanner/reference/ms_generate_map.html)
    function, which automatically produces both `.pdf` and `.png` versions;
 
 2. Either print the `.pdf` version to use as desired in any kind of survey
@@ -88,7 +91,7 @@ The package is designed to enable the following workflow:
    converting it to either `.pdf` or `.png` format; and
 
 5. Rectify the modified version against the original via the
-   [`ms_rectify_maps()`](https://mpadge.github.io/mapscanner/reference/ms_rectify_maps.html)
+   [`ms_rectify_maps()`](https://docs.ropensci.org/mapscanner/reference/ms_rectify_maps.html)
    function, which distinguishes individual annotations,
    and converts each one to a spatial object able to be analysed in any desired
    manner.
@@ -99,7 +102,7 @@ The `mapscanner` package is intended to aid a *practical* workflow, and so
 a few practical tips may be recommended here to ensure best results:
 
 1. The original digital files generated with
-   [`ms_generate_map()`](https://mpadge.github.io/mapscanner/reference/ms_generate_map.html)
+   [`ms_generate_map()`](https://docs.ropensci.org/mapscanner/reference/ms_generate_map.html)
    are necessary to rectify subsequently drawn-on and scanned maps, and so must
    be retained at all times.
 2. Marks drawn on maps should be *coloured* -- any black or grey markings will
@@ -148,7 +151,7 @@ for your current R session, you may load `mapscanner`, and then run
 
 Having obtained and set a [`mapbox`](https://mapbox.com) token as described
 above, the
-[`ms_generate_map()`](https://mpadge.github.io/mapscanner/reference/ms_generate_map.html)
+[`ms_generate_map()`](https://docs.ropensci.org/mapscanner/reference/ms_generate_map.html)
 function can be used to generate printable maps for a specified bounding box in
 both `.pdf` and `.png` formats. Usage is a simple as,
 ```{r generate-fakey, eval = FALSE}
@@ -169,7 +172,7 @@ finer control can be gained by clicking beneath this coordinate window on the
 line which says, "Manually select a different area," which brings a drag-able
 rectangle onto the current screen. The coordinates in the bounding box then
 simply need to be entered in to the `bbox` parameter of
-[`ms_generate_map()`](https://mpadge.github.io/mapscanner/reference/ms_generate_map.html)
+[`ms_generate_map()`](https://docs.ropensci.org/mapscanner/reference/ms_generate_map.html)
 in the order (`xmin`, `ymin`, `xmax`, `ymax`) -- or anti-clockwise from the
 left-hand coordinate.
 
@@ -217,7 +220,7 @@ Standard uses of the package should not need to explicitly access or modify
[…]
