+++ README.Rmd
@@ -11,7 +11,7 @@ knitr::opts_chunk$set(
   out.width = "100%"
 )
 ```
-# mapscanner ![](http://www.textfiles.com/underconstruction/CoColosseumHoop5020underconstruction_blk.gif)
+# mapscanner
 
 <!-- badges: start -->
 [![Travis build status](https://travis-ci.org/ropensci/mapscanner.svg?branch=master)](https://travis-ci.org/ropensci/mapscanner)
