### 2020-03-20 #330 @mpadge
@sckott Thanks for all of the much appreciated editorial assistance with this package. I'm very excited that it's finally crossed the threshold over to rOpenSci. Here's you checklist of items again to explicitly confirm that I've attended to all points:

- [x] *link to the COC from the README*
- [x] *on check, there's a package size warning on my macos that it's 5.2 mb, maybe compressing some images may bring that below 5?* - Rectified as described in [issue #27](https://github.com/ropensci/mapscanner/issues/27)
- [x] *i couldn't try the mapdeck eg because couldn't get the dev versions of it's dependencies to install from source to get required versions - no worries, let's move on* - buzz
- [x] *Transfer the repo to rOpenSci'*
- [x] *Add the rOpenSci footer to the bottom of your README*
- [x] *Fix all links to the GitHub repo to point to the repo under the ropensci organization.*
- [x] *If you already have a pkgdown website, fix its URL to point to https://docs.ropensci.org*
- [x] *in your DESCRIPTION file, include the docs link in the URL field alongside the link to the GitHub repository*
- [ ] *Add a mention of the review in DESCRIPTION via rodev::add_ro_desc().*

That function no longer works, but I gave due [credit to both reviewers](https://github.com/ropensci/mapscanner/commit/0fbb49564f585ad420f99c00c319556678b0cee9) in the DESCRIPTION.

- [x]  Fix any links in badges for CI and coverage to point to the ropensci URL.

For JOSS:

- [ ] Submit to JOSS at https://joss.theoj.org/papers/new, using the rOpenSci GitHub repo URL.

Last step not currently possible "Due to Walking Dead World's End" (as sign on a bar in my town helpfully declares).

---

Other than that, it seems to me that that should be sufficient to consider it successfully onboarded, right? :tada: :man_dancing: Dunno what the status of blog entries and suchlike is in the midst of global restructuring, but whatever can be worked out whenever it might be able to be, i'll be part of it. Thanks!
