### 2020-03-20 #330 @stefaniebutland
Sounds good. @mpadge please see our new [rOpenSci Blog Guidelines for Authors and Editors](https://blogguide.ropensci.org/) for updated instructions for a blog post or tech note.
