### 2020-06-09 4b97a31
pre-install a couple more
