+++ .github/workflows/R-CMD-check.yaml
@@ -20,6 +20,7 @@ jobs:
           sudo apt-get update
           sudo apt-get install libcurl4-openssl-dev libxml2-dev libssl-dev pandoc pandoc-citeproc
           sudo apt-get install libudunits2-dev libgdal-dev
+          sudo apt-get install libmagick++-dev libpoppler-cpp-dev
       - name: Install dependencies
         run: |
           install.packages(c("remotes", "rcmdcheck"))
