### 2020-06-09 e287ac6
pre-render mapshot images
