### 2020-06-09 0ec1e9a
fix mapstyle image render
