### 2020-06-09 70a6521
functions for proj strings
