### 2020-06-10 f4c0d48
fix @mdsumner code for #33, #34
The warnings where still issued on both the crop and stackApply calls -
my guess is because the extent objects did not have any CRS, but as these
are all straigh-up raster ops, it makes no diff if the CRS is just appended
at the end - right?
