### 2020-06-10 a1fe4b0
bundle image files for mapstyles vignette
