+++ README.Rmd
@@ -14,10 +14,15 @@ knitr::opts_chunk$set(
 # mapscanner
 
 <!-- badges: start -->
+[![R build
+status](https://github.com/ropensci/mapscanner/workflows/R-CMD-check/badge.svg)](https://github.com/ropensci/mapscanner/actions?query=workflow%3AR-CMD-check)
 [![Travis build status](https://travis-ci.org/ropensci/mapscanner.svg?branch=master)](https://travis-ci.org/ropensci/mapscanner)
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/ropensci/mapscanner?branch=master&svg=true)](https://ci.appveyor.com/project/mpadge/mapscanner)
 [![codecov](https://codecov.io/gh/ropensci/mapscanner/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/mapscanner)
 [![Project Status: Concept](https://www.repostatus.org/badges/latest/concept.svg)](https://www.repostatus.org/#concept)
+
+[![](https://badges.ropensci.org/330_status.svg)](https://github.com/ropensci/onboarding/issues/330)
+
 <!-- badges: end -->
 
 
