### 2020-06-12 a701e85
update rasterBrick output with new CRS [skip ci]
