### 2020-06-12 7e6cda4
mv vignette fig path
