+++ vignettes/mapscanner.Rmd
@@ -254,9 +254,12 @@ rectification, while the second has two red lines drawn upon it:
 <!--
 <img src="omaha-polygons.png" width="600" height="400"/>
 -->
+
 ```{r omaha-poly-png, echo = FALSE, out.width = "75%", eval = TRUE}
 op <- system.file ("extdata", "omaha-polygons.png", package = "mapscanner")
-knitr::include_graphics (op)
+f <- file.path (getwd (), "omama-polygons.png")
+chk <- file.copy (op, f)
+knitr::include_graphics (f)
 ```
 
 
