### 2020-06-18 e67199a
rm appveyor
