### 2020-06-18 4c2be86
add matrix to gh workflow
