+++ inst/paper.md
@@ -1,0 +1,262 @@
+---
+title: mapscanner
+tags:
+    - maps
+    - spatial
+    - R
+authors:
+    - name: Mark Padgham
+      affiliation: 1
+    - name: Michael D. Sumner
+      affiliation: 2
+affiliations:
+    - name: rOpenSci
+      index: 1
+    - name: Australian Antarctic Division, Hobart, Australia.
+      index: 2
+date: 06 Feb 2020
+bibliography: vignettes/mapscanner.bib
+nocite: |
+  @*
+---
+
+# Summary
+
+`mapscanner` is an **R** package that enables maps to be printed out, drawn on,
+scanned back in, and have the drawn components converted to spatial objects. It
+is intended for use in social surveys and similar endeavours in which printed
+paper maps are preferred over digital screens, and enables marks drawn by hand
+on paper maps to be converted to spatial objects. The use of paper maps is
+often advantageous because (i) they cost much less than computer devices
+otherwise needed for digital equivalents; and (ii) they may in many situations
+be deemed to yield more reliable or accurate results through removing the
+psychological barriers often associated with screen-based surveys [@Toninelli2016].
+
+# Usage
+
+The package is designed to enable the following workflow:
+
+1. Generate a map with the `ms_generate_map()` function, which automatically
+    produces both `.pdf` and `.png` versions;
+2. Either print the `.pdf` version to use as desired in any kind of survey
+   environment, or use either the `.pdf` or `.png` versions in digital form for
+   screen-based surveys.
+3. Draw on the map;
+4. For paper maps, digitise the drawn-on (from here on, "modified") map,
+   converting it to either `.pdf` or `.png` format; and
+5. Rectify the modified version against the original via the
+   `ms_rectify_maps()` function, which distinguishes individual annotations,
+   and converts each one to a spatial object able to be analysed in any
+   desired manner.
+
+The following two sections describe the two primary functions of the
+`mapscanner` package, corresponding to the two primary steps of producing maps
+to be used in surveys (or other activities), and rectifying modified maps
+against these originals in order to extract spatial objects. The second of
+these sections also describes the kinds of markings able to be recognised, and
+the kinds of spatial objects to which these may be converted.
+
+
+## Map generation
+
+Having obtained and set a [`mapbox`](https://mapbox.com) token as described
+in the package website, the `ms_generate_map()` function can be used to
+generate printable maps for a specified bounding box in both `.pdf` and `.png`
+formats. Usage is a simple as,
+``` r
+ms_generate_map ("chennai india", mapname = "chennai")
+```
+
+The two generated maps are saved in the current working directory (`getwd()`).
+To save maps in alternative locations, the `mapname` parameter can optionally
+specify paths. To provide finer control over the scales of maps, precise
+bounding boxes can also be submitted. To determine desired bounding boxes, we
+recommend using the ['openstreetmap.org' website](https://openstreetmap.org),
+zooming to a desired area, then clicking the "Export" button. A window will
+appear which includes the bounding coordinates of the current screen. Even
+finer control can be gained by clicking beneath this coordinate window on the
+line which says, "Manually select a different area," which brings a drag-able
+rectangle onto the current screen. The coordinates in the bounding box then
+simply need to be entered in to the `bbox` parameter of `ms_generate_map()` in
+the order (`xmin`, `ymin`, `xmax`, `ymax`) -- or anti-clockwise from the
+left-hand coordinate.
+
+The amount of detail in resultant maps is controlled by the `max_tiles`
+argument, with larger values producing more detail, and resulting in larger
+file sizes. The default value of `max_tiles = 16L` (where the `L` symbol tells
[…]
