### 2020-06-18 e259b18
add cran-comments [ci skip]
